requirements.txt
astroid==3.3.9
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
dill==0.4.0
filelock==3.18.0
greenlet==3.2.1
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isort==6.0.1
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
jsontemplates==0.1.0
lxml==5.4.0
mccabe==0.7.0
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
platformdirs==4.3.7
playwright==1.51.0
Protego==0.4.0
proxyscrape==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyee==12.1.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dotenv==1.1.0
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
scrapy-playwright==0.0.43
scrapy-proxy-pool==0.1.9
service-identity==24.2.0
setuptools==79.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tldextract==5.3.0
tomlkit==0.13.2
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
typing_extensions==4.13.2
urllib3==2.4.0
w3lib==2.3.1
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
